Move parser.rs -> ast.rs
This makes sense, since the lossless AST is quite large.
